policy: Group policy and identity management cells
Move policy.Cell to policy/repository.Cell, so that policy.Cell can serve as the parent cell that includes cells:
- policy repository
- policy k8s cell
- policy directory
- identity cache (allocator)
- identity manager
